requirements.txt
beautifulsoup4==4.9.3
pycryptodome==3.9.9
requests==2.21.0
PyYAML==5.3.1
